requirements.txt
# Библиотеки ниже при локальном использовании модели
# PyTorch: локальная загрузка - https://pytorch.org/get-started/locally/
# torch==2.9.1 --index-url https://download.pytorch.org/whl/cpu (пример установки)
# transformers==4.57.3
# hf-xet==1.2.0

huggingface-hub==0.36.0
fastapi==0.124.4
uvicorn==0.38.0
python-dotenv==1.2.1
requests==2.32.5
pytest==9.0.2
